product/requirements.txt
flask
python-dotenv
flask-restx
requests
zeep
urllib3
mysql.connector
simple_salesforce
pexpect
paramiko
pyyaml
Jinja2
